@@ -60,7 +60,7 @@ task(TASK_JEST, 'Runs jest tests').setAction(
 )
 module.exports = {
   solidity: {
-    version: '0.8.7',
+    version: '0.8.13',
     settings: {
       optimizer: {
         enabled: true,
