@@ -5,7 +5,7 @@ import keyHoldersByLocks from '../queries/keyholdersByLock'
 import { ToastHelper } from '../components/helpers/toast.helper'
 import keyholdersByKeyIdQuery from '../queries/keyholdersByKeyId'
 import { getValidNumber } from '~/utils/strings'
-import { MemberFilters } from '~/unlockTypes'
+import { MemberFilter } from '~/unlockTypes'
 export class GraphService {
   public client: any
 
@@ -56,14 +56,12 @@ export class GraphService {
   }: {
     locks: string[]
     expireTimestamp: number
-    expiration: MemberFilters
+    expiration: MemberFilter
     first: number
     skip: number
     search: string | number
     filterKey: string
   }) => {
-    const showActive = expiration === MemberFilters.ACTIVE
-
     let query
     const keyId = getValidNumber(search)
 
@@ -71,7 +69,8 @@ export class GraphService {
     if (filterKey === 'keyId' && `${search}`?.length) {
       query = keyholdersByKeyIdQuery()
     } else {
-      query = keyHoldersByLocks(showActive)
+      // pass expiration type filter 'all' | 'expired' | 'active'
+      query = keyHoldersByLocks(expiration)
     }
 
     const owner = `${search}`?.toLowerCase() ?? ''
